chore: refactor survey overview
